Drop dead code, fix native warnings, fix a stale build var
- native_bridge.c / NativeBridge.java: remove the dsv2 native method.
  It was registered in the JNINativeMethod table and declared in Java
  but never called from either side. Dropped both so the registration
  table still matches 1:1.
- IconButtonComponent: remove a stray @deprecated that came out of the
  decompile — the class is actively used by every icon button.
- controller_ui.cpp: the per-frame delta was a double narrowed to float
  at every easeTo call (8 C4244s). Compute as float directly; the
  0..0.1 range is exact in float. Native build is now warning-free.
- build.bat: drop the PATCH variable that pointed at an
  english-localization.patch which does not exist and was never used.

+++ native/loader_gui/controller_ui.cpp
@@ -177,8 +177,9 @@ bool ControllerUi::pointerIn(float x, float y, float width, float height) const 
 
 void ControllerUi::updateFrame() {
     const auto now = std::chrono::steady_clock::now();
-    const double delta = std::clamp(std::chrono::duration<double>(now - lastFrame_).count(),
-        0.0, 0.1);
+    const float delta = static_cast<float>(std::clamp(
+            std::chrono::duration<double>(now - lastFrame_).count(),
+            0.0, 0.1));
     lastFrame_ = now;
     model_.tick();
